Target find bug fixes continue.  Needed to update for others to get going with development.  Changes to be committed: 	modified:   AngleCorrections.c 	modified:   BoardComm.c 	modified:   BoardComm.h 	modified:   DoCoarseScan.c 	modified:   DoCoarseScan.h 	modified:   DoCoarseScan2.c 	modified:   DoCoarseScan2.h 	modified:   DoFindOneTarget.c 	modified:   Init.c 	modified:   LaserCmds.c 	modified:   LaserCmds.h 	modified:   LaserFlex.c 	modified:   LaserInterface.c 	modified:   LaserInterface.h 	modified:   Main.c 	modified:   Makefile 	modified:   Protocol.h 	modified:   QuickCheckManager.c 	modified:   RightOnCert.c 	modified:   SensorSearch.c 	modified:   ShowTargets.c 	modified:   parse_data.c 	modified:   parse_data.h

+++ DoCoarseScan.h
@@ -1,8 +1,7 @@
-#include <stdint.h>
 #ifndef DOCOARSESCAN_H
 #define DOCOARSESCAN_H
 
 int DoCoarseScan(struct lg_master *pLgMaster, int16_t dX, int16_t dY,
-		 uint32_t lsstep, int lscount,
+		 int16_t lsstep, uint16_t lscount,
 		 int16_t *xfound, int16_t *yfound);
 #endif
