+++ DoCoarseScan2.h
@@ -2,7 +2,7 @@
 #define DOCOARSESCAN2_H
 
 int DoCoarseScan2(struct lg_master *pLgMaster, int16_t dX, int16_t dY,
-		  uint32_t lsstep, int lscount, int16_t *xfound,
+		  uint16_t lsstep, uint16_t lscount, int16_t *xfound,
 		  int16_t *yfound);
